fixing issue in grain generator where the inside check in shape ops was returning an int instead of a float and was cause grains to be twice as big as needed

+++ Source/DREAM3D/Common/ShapeOps/CubeOctohedronOps.h
@@ -28,7 +28,7 @@ namespace DREAM3D
 
       virtual float radcur1(std::map<ArgName, float> args);
 
-      virtual int inside(float axis1comp, float axis2comp, float axis3comp);
+      virtual float inside(float axis1comp, float axis2comp, float axis3comp);
       virtual void init() { Gvalue = 0.0f; }
 
     protected:
